refactored configs, better error handling when service is not found
